Abstract core ui into core folder
